fix some bugs. for Categories table user enter category_id instead of Category Name

+++ p2-scape-goats/src/main/java/com/revature/tarotcards/p2scapegoats/controllers/john/JohnConsultantController.java
@@ -2,9 +2,14 @@ package com.revature.tarotcards.p2scapegoats.controllers.john;
 
 
 import com.revature.tarotcards.p2scapegoats.dtos.john.request.JohnNewConsultantRequest;
+import com.revature.tarotcards.p2scapegoats.dtos.john.response.JohnPrincipal;
 import com.revature.tarotcards.p2scapegoats.dtos.request.melissa.NewConsultantRequest;
 import com.revature.tarotcards.p2scapegoats.models.melissa.Consultants;
+import com.revature.tarotcards.p2scapegoats.models.melissa.Roles;
 import com.revature.tarotcards.p2scapegoats.services.john.JohnConsultantService;
+import com.revature.tarotcards.p2scapegoats.services.melissa.RoleService;
+import com.revature.tarotcards.p2scapegoats.services.melissa.TokenService;
+import com.revature.tarotcards.p2scapegoats.utils.melissa.custom_exceptions.AuthenticationException;
 import com.revature.tarotcards.p2scapegoats.utils.melissa.custom_exceptions.InvalidRequestException;
 import com.revature.tarotcards.p2scapegoats.utils.melissa.custom_exceptions.ResourceConflictException;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -21,14 +26,29 @@ public class JohnConsultantController {
     @Autowired
     private JohnConsultantService consultantService;
 
+    @Autowired
+    private TokenService tokenService;
+
+    @Autowired
+    private RoleService roleService;
+
 
     @CrossOrigin
     @ExceptionHandler(value = {ResourceConflictException.class, InvalidRequestException.class})
     @ResponseStatus(value = HttpStatus.BAD_REQUEST)
     @PostMapping(value = "", consumes = "application/json", produces = MediaType.APPLICATION_JSON_VALUE)
-    public @ResponseBody Consultants createConsultant(@RequestBody JohnNewConsultantRequest request) {
+    public @ResponseBody Consultants createConsultant(@RequestHeader(value = "authorization") String token, @RequestBody JohnNewConsultantRequest request) {
+        if (token == null) {
+            throw new AuthenticationException("Sorry, you are not authorized to make this request");
+        }
 
-        return consultantService.save(request);
+        JohnPrincipal principal = tokenService.extractRequesterDetails(token);
+        Roles role = roleService.findByRole_id(principal.getRole());
+        if (role.getTitle().equalsIgnoreCase("USERS") || role.getTitle().equalsIgnoreCase("ADMIN"));
+        {
+            return consultantService.save(request);
+        }
+
     }
 
 
@@ -36,8 +56,19 @@ public class JohnConsultantController {
     @ExceptionHandler(value = {ResourceConflictException.class, InvalidRequestException.class})
     @ResponseStatus(value = HttpStatus.BAD_REQUEST)
     @PutMapping(value = "", consumes = "application/json", produces = MediaType.APPLICATION_JSON_VALUE)
-    public @ResponseBody Consultants updateConsultant(@RequestBody JohnNewConsultantRequest request) {
-        return consultantService.update(request);
+    public @ResponseBody Consultants updateConsultant(@RequestHeader(value = "authorization") String token, @RequestBody JohnNewConsultantRequest request) {
+
+        if (token == null) {
+            throw new AuthenticationException("Sorry, you are not authorized to make this request");
+        }
+
+        JohnPrincipal principal = tokenService.extractRequesterDetails(token);
+        Roles role = roleService.findByRole_id(principal.getRole());
+        if (role.getTitle().equalsIgnoreCase("USERS") || role.getTitle().equalsIgnoreCase("ADMIN"));
+        {
+            return consultantService.update(request);
+        }
+
     }
 
 
@@ -46,16 +77,35 @@ public class JohnConsultantController {
     @ExceptionHandler(value = {ResourceConflictException.class, InvalidRequestException.class})
     @ResponseStatus(value = HttpStatus.BAD_REQUEST)
     @GetMapping(value="", produces = MediaType.APPLICATION_JSON_VALUE)
-    public @ResponseBody List<Consultants> getAll() {
-        return consultantService.getAllConsultants();
+    public @ResponseBody List<Consultants> getAll(@RequestHeader(value = "authorization") String token) {
+        if (token == null) {
+            throw new AuthenticationException("Sorry, you are not authorized to make this request");
+        }
+
+        JohnPrincipal principal = tokenService.extractRequesterDetails(token);
+        Roles role = roleService.findByRole_id(principal.getRole());
+        if (role.getTitle().equalsIgnoreCase("USERS") || role.getTitle().equalsIgnoreCase("ADMIN"));
+        {
+            return consultantService.getAllConsultants();
+        }
+
     }
 
     @CrossOrigin
     @ExceptionHandler(value = {ResourceConflictException.class, InvalidRequestException.class})
     @ResponseStatus(value = HttpStatus.BAD_REQUEST)
     @DeleteMapping(value = "", consumes = "application/json", produces = MediaType.APPLICATION_JSON_VALUE)
-    public @ResponseBody String deleteCategory(@RequestBody JohnNewConsultantRequest request) {
-        consultantService.deleteConsultant(request);
+    public @ResponseBody String deleteConsultant(@RequestHeader(value = "authorization") String token, @RequestBody JohnNewConsultantRequest request) {
+        if (token == null) {
+            throw new AuthenticationException("Sorry, you are not authorized to make this request");
+        }
+
+        JohnPrincipal principal = tokenService.extractRequesterDetails(token);
+        Roles role = roleService.findByRole_id(principal.getRole());
+        if (role.getTitle().equalsIgnoreCase("USERS") || role.getTitle().equalsIgnoreCase("ADMIN"));
+        {
+            consultantService.deleteConsultant(request);
+        }
         return "Deleted : " +  request.getId();
     }
